// src/types/theme.d.ts
export type ThemeType = {
    backgroundColor: string;
    gradientColors: string[];
    gradientMouseColor: string[];
    color: string;
    CardTranparente: string;
    BackgroundNavbar: string;
    ActiveButtonNav: string;
    // Otros colores y propiedades que necesites
    CardData: string;
    CardHover: string;
    SecundaryText: string;
    CompanyText: string;
    DescripcionCompanyText: string,
    BotonRun: string,
    FondoGit: string,
    cardGitTransparente: string
  };
  